+++ javaparser-core-testing/src/test/java/com/github/javaparser/printer/lexicalpreservation/Issue2374Test.java
@@ -27,11 +27,14 @@ import com.github.javaparser.StaticJavaParser;
 import com.github.javaparser.ast.body.MethodDeclaration;
 import com.github.javaparser.ast.stmt.Statement;
 import java.util.Optional;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 public class Issue2374Test extends AbstractLexicalPreservingTest {
 
     @Test
+    @Disabled(
+            "Test disabled because this test fails on Ubuntu systems regardless of the Java version, but works on Windows and macOS systems")
     public void test() {
         String lineComment = "Example comment";
         considerCode("public class Bar {\n" + "    public void foo() {\n"